adding test for DEER with specified climate zone
Tests passed, I think issue in URBANopt workflow may have had use_upstream_args to do and pulled climate zone from ChangeBuildingLocation which may have been "T24-CEC5" instead of "CEC T24-CEC5" which is used in standards. It is a bit awkward that they are not the same. That isn't the case for ASHRAE climate zones